#Metatavu/pyramus#1 - Renamed domainmodel package
