+++ persistence/src/main/java/fi/otavanopisto/pyramus/domainmodel/matriculation/MatriculationExamAttendanceStatus.java
@@ -1,4 +1,4 @@
-package fi.otavanopisto.pyramus.domainmodel.matriculationexamapplication;
+package fi.otavanopisto.pyramus.domainmodel.matriculation;
 
 public enum MatriculationExamAttendanceStatus {
   PLANNED,
