+++ persistence/src/main/java/fi/otavanopisto/pyramus/domainmodel/matriculation/MatriculationExamSubject.java
@@ -1,4 +1,4 @@
-package fi.otavanopisto.pyramus.domainmodel.matriculationexamapplication;
+package fi.otavanopisto.pyramus.domainmodel.matriculation;
 
 public enum MatriculationExamSubject {
   MAA,
