+++ persistence/src/main/java/fi/otavanopisto/pyramus/domainmodel/matriculation/MatriculationExamTerm.java
@@ -1,0 +1,6 @@
+package fi.otavanopisto.pyramus.domainmodel.matriculation;
+
+public enum MatriculationExamTerm {
+  SPRING,
+  AUTUMN
+}
