+++ persistence/src/main/java/fi/otavanopisto/pyramus/domainmodel/matriculation/SchoolType.java
@@ -1,0 +1,6 @@
+package fi.otavanopisto.pyramus.domainmodel.matriculation;
+
+public enum SchoolType {
+  UPPERSECONDARY,
+  VOCATIONAL;
+}
