--- persistence/src/main/java/fi/otavanopisto/pyramus/domainmodel/matriculationexamapplication/MatriculationExamTerm.java
@@ -1,6 +1,0 @@
-package fi.otavanopisto.pyramus.domainmodel.matriculationexamapplication;
-
-public enum MatriculationExamTerm {
-  SPRING,
-  AUTUMN
-}
