--- persistence/src/main/java/fi/otavanopisto/pyramus/domainmodel/matriculationexamapplication/SchoolType.java
@@ -1,6 +1,0 @@
-package fi.otavanopisto.pyramus.domainmodel.matriculationexamapplication;
-
-public enum SchoolType {
-  UPPERSECONDARY,
-  VOCATIONAL;
-}
